Initial rework of the ThumbGolf handler model.
Veneering is now done unconditionally. This is slower, but it greatly
simplifies things overall. We no longer need to go through hoops to
call functions that are not Async-Signal-Safe.

We also use a new handler model which adds a "busy" state.

Additionally, it uses a state struct which lets us access the CPSR
using a bitfield.

It still needs some major cleanup though, and the narrow I/O functions
are designed for the old model and must be updated.